# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Very Useful Tools to Remember",
    "description": "API of to manage your tools information like names, links, description and tags",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local server"
    }
  ],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/tools": {
      "get": {
        "tags": ["Tools"],
        "summary": "Gets tools",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/tools"
            }
          },
          "400": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": ["Tools"],
        "summary": "Create a new tool",
        "requestBody": {
          "description": "Tool Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/tool"
              }
            }
          }
        },
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/id"
            }
          },
          "400": {
            "description": ""
          }
        }
      }
    },
    "/tools/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of the tool that we want to match",
          "type": "string"
        }
      ],
      "get": {
        "tags": ["Tools"],
        "summary": "Gets a tool given ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Tool with id",
            "schema": {
              "$ref": "#/definitions/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/tools"
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "summary": "Delete a tool with given ID",
        "tags": ["Tools"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Tool ID",
            "schema": {
              "$ref": "#/definitions/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "id": {
      "properties": {
        "uuid": {
          "type": "string"
        }
      }
    },
    "tool": {
      "type": "object",
      "properties": {
        "id": {
          "$ref": "#/definitions/id"
        },
        "title": {
          "type": "string"
        },
        "link": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "tools": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/tool"
      }
    }
  }
}